carvers/requirements.txt
newspaper3k
selenium
beautifulsoup4
Pillow
cloudscraper
pandas
ssdeep
gunicorn
flask
